pyproject.toml
[tool.poetry]
name = "taxi-service"
version = "0.0.0"
description = "Taxi Grid Service"
authors = ["Artem Asel Dima"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
redis = {extras = ["hiredis"], version = "^5.0.1"}
pydantic = {extras = ["email"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
asyncpg = "^0.30.0"
python-json-logger = "^4.0.0"
PyJWT = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
fakeredis = {extras = ["aioredis"], version = "^2.20.0"}
httpx = "^0.25.1"
pytest-dotenv = "^0.5.2"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"